reverting back to trying md5 for image diffs first
if images are not identical (hashes are not identical) then we can fall
back to a detailed comparison procedure.

i was initially under the impression that the detailed comparison loop
took the most time. However, after running some timers, it's clear that
the bottleneck is reading the images into ChunkyPNG, frequently taking
close to a minute.